refactor/extracted classes from multi class
